+++ src/rectangle.c
@@ -347,7 +347,7 @@ int RTreeContained(struct Rect *R, struct Rect *S)
 {
     register struct Rect *r = R, *s = S;
     register int i, j, result;
-    assert((int)r && (int)s);
+    assert((long)r && (long)s);
 
      /* undefined rect is contained in any other */
     if (Undefined(r))
